requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.11
pydantic==1.10.12
pytest==7.4.0
requests==2.31.0
